Boy Walking
At this point the boy moves with the left and right arrow. They move
from Denise’s avatar_control. The jump_avatar file makes implements a
transformer to check if the avatar is on the ground, if not a blend
tree is implemented (this means that the avatar will change from
walking to the jump position.) Still need to implement the falling, and
off of a button click. In scene called MockLevel.