Added new enableLineDrawingTestLvl and DrawPhysicsLineTestLvl scripts
These scripts are the counterparts to the scripts enableLineDrawing and
DrawPhysicsLine (the scripts for drawing zip lines specifically in
Level 1 Scene ONLY). These new scripts labeled with “TestLvl” are to be
used in the Test Level Scene only - the width of the zip line segments
have to be different in the different scenes because of screen
resolution and such.