Positioned all game objects in correct z plane (not layers)
The following interactive game objects (snap
points and zip line segments; later power up attachments and landing
platforms) now have transform z position set to 0. Transform z position
of BoyZipping is set to -1 to appear in front of the snap points.
Transforms of furniture
game objects (these GO’s the player doesn’t interact with) have z
positions in the range of z=1 to 4 (this is how we stagger the
furniture in front and behind each other, instead of creating layers).