Can erase line by destroying line GO on mouse down, but still somehow not restricted to just erase mode
Changed Line GO box collider from a child of the line GO to a directly
attached component of the line GO. This allows the OnMouseDown() to
work and the line GO is destroyed when the. NOT QUITE RIGHT: you can
destroy the line GO *anytime* you click on a line GO, but we want to
restrict this ability entirely to erase mode (activated by
enableLineErasing.cs script).