Curves Additions to Boy_3_1
to clarify the naming, this is a male avatar with outfit 3 skin 1.
(shark shirt, fair skin).

I added components to the avatar: rigid body

created a new folder under animating where I will store the animations.
They will be named with the name of the avatar followed by what they
are doing.