Added enableRebuild object with enableRebuild script attached to implement RebuildButton in Canvas
When you press the RebuildButton, ResetAvatar() is called to return the
avatar to build mode. This includes resetting the position of the
avatar to its initial position (before pressing play) and destroys the
avatar’s rigidbody component (so that the avatar won’t interact with
the Unity physics during build mode).